Combine Wind struct model

+++ WeatherAPP/WeatherAPP/Model/Weather.swift
@@ -71,5 +71,5 @@ struct Weather: Codable {
 // MARK: - Wind
 struct Wind: Codable {
     let speed: Double
-    let deg: Int
+    let deg: Double
 }
